Move db prepair functions
