+++ create_tables.py
@@ -1,6 +1,6 @@
 import asyncio
 import logging
-from app.database.base import create_tables, drop_tables
+from app.database.utils import create_tables, drop_tables
 
 logging.basicConfig(level=logging.INFO)
